Imported a bunch of art from thoseawesomeguys.com/prompts, reworked data tree.
This data is public domain (Creative Commons 0, or "CC0"), used in 100+ games,
and quite high quality! It covers just about all the popular controllers,
so we're hitting the ground running now!

Data came from https://thoseawesomeguys.com/prompts/ , originally in FLA
format. Bekoha, from Twitter, was gracious enough to convert it to SVG format
for me, including taking the time to figure out _how_ to convert it.

Note that this might need tweaking; I had to guess which paddle buttons on the
Steam Deck map to which SDL gamepad button, etc, but I think overall the data
set is mostly correct.

Fixes #4.